ajuste de codigo y script para deployar con docker

+++ BaseDatos.sql
@@ -6,7 +6,7 @@ DROP TABLE IF EXISTS cuenta;
 DROP TABLE IF EXISTS cliente;
 
 CREATE TABLE cliente (
-    id INT NOT NULL AUTO_INCREMENT,
+    id BIGINT NOT NULL AUTO_INCREMENT,
     nombre VARCHAR(100) NOT NULL,
     genero VARCHAR(20),
     edad INT,
@@ -20,23 +20,23 @@ CREATE TABLE cliente (
 );
 
 CREATE TABLE cuenta (
-    id INT NOT NULL AUTO_INCREMENT,
+    id BIGINT NOT NULL AUTO_INCREMENT,
     numero_cuenta VARCHAR(20) UNIQUE NOT NULL,
     tipo_cuenta VARCHAR(50) NOT NULL,
     saldo_inicial DECIMAL(15,2) NOT NULL DEFAULT 0.00,
     estado BOOLEAN NOT NULL DEFAULT TRUE,
-    cliente_id INT NOT NULL,
+    cliente_id BIGINT NOT NULL,
     PRIMARY KEY (id),
     CONSTRAINT fk_cliente_cuenta FOREIGN KEY (cliente_id) REFERENCES cliente(id) ON DELETE CASCADE
 );
 
 CREATE TABLE movimiento (
-    id INT NOT NULL AUTO_INCREMENT,
+    id BIGINT NOT NULL AUTO_INCREMENT,
     fecha DATETIME NOT NULL DEFAULT CURRENT_TIMESTAMP,
     tipo_movimiento VARCHAR(50) NOT NULL,
     valor DECIMAL(15,2) NOT NULL,
     saldo DECIMAL(15,2) NOT NULL,
-    cuenta_id INT NOT NULL,
+    cuenta_id BIGINT NOT NULL,
     PRIMARY KEY (id),
     CONSTRAINT fk_cuenta_movimiento FOREIGN KEY (cuenta_id) REFERENCES cuenta(id) ON DELETE CASCADE
 );
@@ -54,7 +54,7 @@ VALUES (NOW(), 'Deposito', 500.00, 2500.00, 1);
 DELIMITER $$
 
 CREATE DEFINER=root@localhost PROCEDURE sp_generar_reporte(
-    IN p_cliente_id INT,
+    IN p_cliente_id BIGINT,
     IN p_fecha_inicio DATETIME,
     IN p_fecha_fin DATETIME
 )
